+++ src/main/java/chess/domain/game/BoardFactory.java
@@ -1,9 +1,9 @@
 package chess.domain.game;
 
 import chess.domain.piece.*;
-import chess.domain.position.Column;
-import chess.domain.position.Position;
-import chess.domain.position.Row;
+import chess.domain.location.Column;
+import chess.domain.location.Position;
+import chess.domain.location.Row;
 
 import java.util.HashMap;
 import java.util.Map;
