+++ src/main/java/chess/domain/location/Row.java
@@ -1,4 +1,4 @@
-package chess.domain.position;
+package chess.domain.location;
 
 public enum Row {
     ONE("1"),
