+++ src/main/java/chess/domain/moveStrategy/PawnSingleMove.java
@@ -1,7 +1,7 @@
 package chess.domain.moveStrategy;
 
 import chess.domain.piece.Color;
-import chess.domain.piece.Vector;
+import chess.domain.location.Vector;
 
 import java.util.Collections;
 import java.util.List;
