+++ src/main/java/chess/domain/piece/Division.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.position.Position;
+import chess.domain.location.Position;
 
 import java.util.List;
