+++ src/main/java/chess/domain/piece/Rook.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.moveStrategy.MoveStrategy;
 import chess.domain.moveStrategy.RookMove;
-import chess.domain.position.Position;
+import chess.domain.location.Position;
 
 import java.util.List;
