+++ src/test/java/chess/domain/game/EmptyBoardMap.java
@@ -2,7 +2,7 @@ package chess.domain.game;
 
 import chess.domain.piece.Empty;
 import chess.domain.piece.Piece;
-import chess.domain.position.Position;
+import chess.domain.location.Position;
 
 import java.util.HashMap;
 import java.util.Map;
