+++ src/test/java/chess/domain/location/PositionTest.java
@@ -1,6 +1,5 @@
-package chess.domain.position;
+package chess.domain.location;
 
-import chess.domain.piece.Vector;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
