+++ src/test/java/chess/domain/location/RowTest.java
@@ -1,4 +1,4 @@
-package chess.domain.position;
+package chess.domain.location;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
